requirements.txt
torch
torchvision
torchaudio
Pillow
datasets
transformers
accelerate
bitsandbytes
open_clip_torch
tiktoken
lightning==2.1.3
hydra-core
wandb
opencv-python
peft
deepspeed
einops
xformers
rich
timm
flash-attn
deepspeed
### Below requirements are more optional and necessary for some evaluations (mainly CIDEr)
sentence-transformers
pycocoevalcap
python-dotenv
spacy
pythainlp
sudachipy
sudachidict_core
nltk